Clarified license terms and made them make sense
Previously all files were released under the terms of the LaTeX public
license, which means altered versions of the files have to be annotated
with what changes were made by who. Since the whole point of this
template is to modify it, that didn't really make so much sense.

So, now only the ucl_thesis.cls file (which shouldn't be commonly
modified) is under the LaTeX public license, and everything else is CC-0
so people can do whatever they want with it.

This is explained in the README.md.